refactor: split remaining oversized files across docs, index, daemon, logging, config
Split daemon/ipc.ts (428→16), logging/index.ts (546→89),
config/loading.ts (420→303), docs/docs-map-generator.ts (447→334),
index/fts5.ts (469→377). Add eslint-disable to logging/types.ts
(types-only file). Add core/internal exclusion for docs package.

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
